requirements.txt
fastapi
uvicorn
pydantic-settings
python-multipart
httpx